# docs/swagger.yaml
basePath: /
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  github_com_simesaba80_toybox-back_internal_interface_schema.CreateTagInput:
    properties:
      name:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  github_com_simesaba80_toybox-back_internal_interface_schema.GetCurrentUserResponse:
    properties:
      display_name:
        type: string
      icon_url:
        type: string
      id:
        type: string
    type: object
  github_com_simesaba80_toybox-back_internal_interface_schema.GetDiscordAuthURLResponse:
    properties:
      url:
        type: string
    type: object
  github_com_simesaba80_toybox-back_internal_interface_schema.GetDiscordTokenResponse:
    properties:
      access_token:
        type: string
    type: object
  github_com_simesaba80_toybox-back_internal_interface_schema.GetUserOutput:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      github_id:
        type: string
      id:
        type: string
      name:
        type: string
      profile:
        type: string
      updated_at:
        type: string
      x_username:
        type: string
    type: object
...
  github_com_simesaba80_toybox-back_internal_interface_schema.TagDetailResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      work_count:
        type: integer
    type: object
  github_com_simesaba80_toybox-back_internal_interface_schema.TagListResponse:
    properties:
      tags:
        items:
          $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.TagDetailResponse'
        type: array
    type: object
  github_com_simesaba80_toybox-back_internal_interface_schema.TagResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  github_com_simesaba80_toybox-back_internal_interface_schema.UpdateUserInput:
    properties:
      display_name:
        maxLength: 32
        minLength: 1
        type: string
      github_id:
        type: string
      profile:
        maxLength: 500
        type: string
      x_username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API server for the Toybox application.
  title: Toybox API
  version: "1.0"
paths:
...
  /auth/discord:
    get:
      description: Get Discord authentication URL
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.GetDiscordAuthURLResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get Discord authentication URL
      tags:
      - auth
  /auth/discord/callback:
    get:
      description: Get Discord token by code
      parameters:
      - description: Discord code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.GetDiscordTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get Discord token by code
      tags:
      - auth
  /auth/logout:
    post:
      description: Logout user and invalidate refresh token
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Logout user
      tags:
      - auth
  /auth/refresh:
    post:
      description: Regenerate token(need refresh token in cookie)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.GetDiscordTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Regenerate token
      tags:
      - auth
  /auth/tags:
    post:
      consumes:
      - application/json
      description: Create a new tag (authentication required)
      parameters:
      - description: Tag to create
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.CreateTagInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.TagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Create a new tag
      tags:
      - tags
  /auth/users:
    patch:
      consumes:
      - application/json
      description: Partially update a user
      parameters:
      - description: Fields to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.GetUserOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
  /auth/users/me:
    get:
      description: Get the current authenticated user's ID, display name, and icon
        URL
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.GetCurrentUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Get the current authenticated user
      tags:
      - users
...
  /tags:
    get:
      description: Get all tags
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.TagListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get all tags
      tags:
      - tags
...
  /users/{id}:
    get:
      description: Get a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.GetUserOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get a user by ID
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Bearer token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"