# internal/interface/schema/user.go
package schema

import (
	"time"

	"github.com/google/uuid"
	"github.com/simesaba80/toybox-back/internal/domain/entity"
)

type GetUserOutput struct {
	ID          uuid.UUID `json:"id"`
	Name        string    `json:"name"`
	DisplayName string    `json:"display_name"`
	Profile     string    `json:"profile"`
	AvatarURL   string    `json:"avatar_url"`
	XUsername   string    `json:"x_username"`
	GithubID    string    `json:"github_id"`
	CreatedAt   string    `json:"created_at"`
	UpdatedAt   string    `json:"updated_at"`
}

type GetCurrentUserResponse struct {
	ID          uuid.UUID `json:"id"`
	DisplayName string    `json:"display_name"`
	IconURL     string    `json:"icon_url"`
}

type UserListResponse struct {
	Users []GetUserOutput `json:"users"`
}

type UpdateUserInput struct {
	DisplayName *string `json:"display_name,omitempty" validate:"omitempty,min=1,max=32"`
	Profile     *string `json:"profile,omitempty" validate:"omitempty,max=500"`
	XUsername   *string `json:"x_username,omitempty" validate:"omitempty"`
	GithubID    *string `json:"github_id,omitempty" validate:"omitempty"`
}

func ToUserResponse(user *entity.User) GetUserOutput {
	if user == nil {
		return GetUserOutput{}
	}
	return GetUserOutput{
		ID:          user.ID,
		Name:        user.Name,
		DisplayName: user.DisplayName,
		Profile:     user.Profile,
		AvatarURL:   user.AvatarURL,
		XUsername:   user.XUsername,
		GithubID:    user.GithubID,
		CreatedAt:   user.CreatedAt.Format(time.RFC3339),
		UpdatedAt:   user.UpdatedAt.Format(time.RFC3339),
	}
}

func ToCurrentUserResponse(user *entity.User) GetCurrentUserResponse {
	if user == nil {
		return GetCurrentUserResponse{}
	}
	return GetCurrentUserResponse{
		ID:          user.ID,
		DisplayName: user.DisplayName,
		IconURL:     user.AvatarURL,
	}
}